+++ SimpleBlog.Application/CommandHandlers/UserCommandHandlers/DeleteUserCommandHandler.cs
@@ -2,6 +2,7 @@
 using SimpleBlog.Application.Commands.UserCommand;
 using SimpleBlog.Domain.Interfaces.Base;
 using SimpleBlog.Domain.Models;
+using System.Security.Authentication;
 
 namespace SimpleBlog.Application.CommandHandlers.UserCommandHandlers;
 
@@ -16,6 +17,9 @@ public class DeleteUserCommandHandler(IUnitOfWork unitOfWork) : IRequestHandler<
 
         var user =
             _unitOfWork.UserRepository.GetById(request.Id) ?? throw new InvalidOperationException("Usuário não encontrado.");
+
+        if (!user.Id.ToString().Equals(request.UserLoggedInId))
+            throw new InvalidCredentialException("Apenas o próprio usuário pode apagar sua conta.");
 
         _unitOfWork.UserRepository.Delete(user);
